Fix writing NaN to obj file
I have some cases where i see NaN written to the obj file. This is
obviously invalid and causes issues.  I've therefore changed the code to
write 0 out instead.  A better fix would be to work out what causes the
NaN's in the first place but this will do for now.